python-core/requirements.txt

fastapi
uvicorn
python-multipart
openai
elevenlabs
ffmpeg-python
python-dotenv
httpx
requests
playwright
